+++ app/_layout.tsx
@@ -13,7 +13,7 @@ import {
   DefaultTheme,
   ThemeProvider,
 } from '@react-navigation/native';
-import { Stack } from 'expo-router';
+import { router, Stack } from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
 import { StatusBar } from 'expo-status-bar';
 import * as Updates from 'expo-updates';
@@ -23,7 +23,7 @@ import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 import ChangeLogModal from '@/components/ChangeLogsModal';
 import { useColorScheme } from '@/hooks/useColorScheme';
-import { currentVersion } from '@/jotai/atoms';
+import { currentVersion, finichedTutorial } from '@/jotai/atoms';
 import { getAppVersion } from '@/utils';
 
 import changeLogsJSON from '../assets/changelogs.json';
@@ -111,6 +111,7 @@ export default function RootLayout() {
               onClose={() => setShowChangeLogsModal(false)}
             />
           </Suspense>
+
           <StatusBar style="auto" />
           <ThemeProvider
             value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}
@@ -128,6 +129,12 @@ export default function RootLayout() {
                 name="navigation"
                 options={{
                   title: 'تنقل',
+                }}
+              />
+              <Stack.Screen
+                name="tutorial"
+                options={{
+                  title: 'جولة تعليمة',
                 }}
               />
             </Stack>